refactor: prepare multicommand setup
